2025/10/09 chore: unify Clang compile options and fix strict warnings
- aligned all modules to strict Clang warning rules (-Wall -Wextra -Wpedantic)
- fixed shadow, sign-compare, old-style-cast, and reorder warnings
- replaced C-style casts with reinterpret_cast / static_cast
- added [[maybe_unused]] for intentional unused symbols
- cleaned CMakeLists for consistent warning policy across subprojects

+++ apps/order_server/infra/db/OrderRepository.cpp
@@ -262,7 +262,7 @@ OrderRepository::Record OrderRepository::ParseSingle(sql::ResultSet* rs) const {
     r.userId       = safeString("user_id");
     r.productId    = safeString("product_id");
     r.quantity     = rs->getInt("quantity");
-    r.totalAmount  = rs->getDouble("total_amount");
+    r.totalAmount  = static_cast<double>(rs->getDouble("total_amount"));
     r.currency     = safeString("currency");
     r.status       = OrderStatusFromString(safeString("status"));
     r.statusReason = safeString("status_reason");
